split source and destination config for RR's, everything else is just confusing
This is because now, with support for label selectors, we need explicit naming on the destination
side, which means the old idea of "the same thing is done on both sides to figure out the names on
origin and destination" doesn't apply anymore, because the logic is now slightly different between
both sides. To not ack this in the CRD will just create endless confusion.

On-behalf-of: @SAP christoph.mewes@sap.com